refactor(web): single-source the manage group dialog across teacher views
